% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcensmix.R
\docType{data}
\name{blood}
\alias{blood}
\title{Blood pressure of mine workers.}
\format{A data frame with 495 rows and 3 variables

  \itemize{

  \item Systolic.BP. Systolic blood pressure \item Diastolic.BP. Diastolic
  blood pressure \item Amplitude. Amplitude

  }}
\usage{
blood
}
\description{
A dataframe containing the blood pressure of a population of mine workers in
Ghana.
}
\references{
Boehning, D., (2000). Computer-assisted Analysis of Mixtures and
Applications: Meta-analysis, Disease mapping and Others. CRC press.

Gunga, H.C., Forson, K., Amegby, N. and Kirsch, K., (1991). Lebensbedingungen
und gesundheitszustand von berg-und fabrikarbeitern im tropischen regenwald
von Ghana. Arbeitsmedizin Sozialmedizin Praventivmediz in, 17-25.
}
\keyword{datasets}
