% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcensmix.R
\docType{data}
\name{insulate}
\alias{insulate}
\title{Failure times of insulating fluid.}
\format{A vector of length 19.}
\usage{
insulate
}
\description{
A dataset containing measurements of 19 failure times (in minutes) for an
insulating fluid between two electrodes subject to a voltage of 34 KV, see
Nelson (1982).
}
\references{
Nelson, W., (1982). Applied Life Data Analysis. Wiley, New York.
}
\keyword{datasets}
