% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcensmix.R
\name{pcensmix-internal}
\alias{pcensmix-internal}
\alias{.f_u}
\title{Internal function for pcensmix}
\usage{
.f_u(eps, Pdat, muhat1, muhat2, sigmahat1, sigmahat2, pi_hat, muhat1_est,
  muhat2_est, sigmahat1_est, sigmahat2_est, pi_hat_est, ll, W, Z, INERiter, N)
}
\description{
.f_u function is for internal usage only and is not intended to be called
by the user.
}
\author{
Lida Fallah, John Hinde

Maintainer: Lida Fallah <l.fallah22@gmail.com>
}
\keyword{internal}
