% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcensmix.R
\docType{package}
\name{pcensmix-package}
\alias{pcensmix-package}
\title{Generate and fit a model to censored mixture data}
\description{
This package provides functions to generate two-component
  mixture data from various different mixture distribution, generate
  progressive Type-II censored data in a mixture structure and fit a normal
  mixture model using a constrained EM algorithm. In addition, it can create a
  progressive Type-II censored version of a given real dataset and fit a
  normal mixture model to.  Main functions are \link[pcensmix]{pcgen},
  \link[pcensmix]{pcensmixSim} and \link[pcensmix]{pcensmixR}. Example datasets are included for
  illustration.
}
\details{
Package: pcensmix

Type: Package

Version: 1.2-1

Date: 2017-07-24

License: GPL (>= 2)
}
\references{
Aitkin, M., Francis, B., Hinde, J. and Darnell, R., (2009).
Statistical Modelling in R. Oxford: Oxford University Press.

Balakrishnan, N. and Aggarwala, R., (2000). Progressive Censoring: Theory,
Methods, and Applications. Springer Science & Business Media.

Hathaway, R.J., (1985). A constrained formulation of maximum-likelihood
estimation for normal mixture distributions. The Annals of Statistics,
795-800.

McLachlan, G. and Krishnan, T., (2007). The EM Algorithm and Extensions. John
Wiley & Sons.

McLachlan, G. and Peel, D., (2004). Finite Mixture Models. John Wiley & Sons.
}
\author{
Lida Fallah <l.fallah22@gmail.com> and John Hinde

Maintainer: Lida Fallah <l.fallah22@gmail.com>
}
\keyword{package}
