% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcensmix.R
\name{print.pcgen}
\alias{print.pcgen}
\title{Print Method for pcgen Objects}
\usage{
\method{print}{pcgen}(x, ...)
}
\arguments{
\item{x}{object of class \code{pcgen}.}

\item{...}{optional arguments to pass by.}
}
\value{
This function uses the generic function \code{\link[base]{print}}
to print the dataset of class \code{"pcgen"} in a nice format.
}
\description{
This function prints the progressive censored data generated by the S3
class \code{\link{pcgen}}.
}
\examples{
## Generate a two component normal mixture data,
Pdat <- pcgen(r = 80, p = 0.3, data = mixgen(N = 100, dist1 = 'norm',
              dist2 = 'norm', control = list(12, 2, 14, 4, 0.3)))
# and print it.
print(Pdat)

}
